+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - classical and cyclic jacobi eigenv* finder
 - hessenberg tranformer with givens rotations
 - householder hessenberg transformer
+- qr without shift method
 
 ## 0.4.4 - 2017-05-28
 ### Added